fix: font style issue, refactor to grid
